+++ MiDAS/R/MiDAS_summarise.R
@@ -4,7 +4,7 @@
 #'
 #' @inheritParams checkHlaCallsFormat
 #' @inheritParams hlaToAAVariation
-#' @param pos String specifying gene and amino acid position, example
+#' @param aa_pos String specifying gene and amio acid position, example
 #'   \code{"A_9"}.
 #' @param aln Matrix containing amino acid sequence alignments as returned by
 #'   \code{\link{readHlaAlignments}} function. By default function will use
@@ -28,13 +28,13 @@
 #'
 #' @export
 summariseAAPosition <- function(hla_calls,
-                                pos,
+                                aa_pos,
                                 aln = NULL,
                                 na.rm = FALSE) {
   assert_that(
     checkHlaCallsFormat(hla_calls),
-    is.string(pos),
-    see_if(grepl("^[A-Z]+[0-9]*_[0-9]+$", pos),
+    is.string(aa_pos),
+    see_if(grepl("^[A-Z]+[0-9]*_[0-9]+$", aa_pos),
            msg = "amino acid position should be formatted like: A_9."
     ),
     see_if(
@@ -44,8 +44,8 @@ summariseAAPosition <- function(hla_calls,
     isTRUEorFALSE(na.rm)
   )
 
-  gene <- gsub("[0-9]+$", "", pos)
-  pos <- as.integer(gsub(".*_([0-9]+)$", "\\1", pos))
+  gene <- gsub("[0-9]+$", "", aa_pos)
+  aa_pos <- as.integer(gsub(".*_([0-9]+)$", "\\1", aa_pos))
 
   alleles <- dplyr::select(hla_calls, dplyr::starts_with(gene)) %>%
     unlist()
@@ -67,8 +67,8 @@ summariseAAPosition <- function(hla_calls,
 
   assert_that(
     see_if(
-      pos <= ncol(aln),
-      msg = sprintf("amino acid position %i is higher than amino acid sequence length.", pos)
+      aa_pos <= ncol(aln),
+      msg = sprintf("amino acid position %i is higher than amino acid sequence length.", aa_pos)
     ),
     see_if(
       all(i <- alleles_wo_na %in% rownames(aln)),
@@ -81,17 +81,17 @@ summariseAAPosition <- function(hla_calls,
 
   aa <- data.frame(
     allele = gsub("[A-Z]+[0-9]*", "", alleles),
-    residue = aln[alleles_wo_na, pos],
+    residue = aln[alleles_wo_na, aa_pos],
     stringsAsFactors = FALSE
   ) %>%
     group_by(.data$residue) %>%
     summarise(
-      allele = paste(unique(allele), collapse = ", "),
+      allele = paste(unique(.data$allele), collapse = ", "),
       count = n(),
       frequency = percent(.data$count / length(alleles))
     ) %>%
     dplyr::select(
-      !! sprintf("HLA-%s (%i)", gsub("_", "", gene), pos) := .data$residue,
+      !! sprintf("HLA-%s (%i)", gsub("_", "", gene), aa_pos) := .data$residue,
       !! sprintf("HLA-%s alleles", gsub("_", "", gene)) := .data$allele,
       .data$count,
       .data$frequency